OPEN - task 2: Add concept of liberties. http://github.com/cgawron/Rango/issues/issue/2

+++ src/de/cgawron/go/Neighborhood.java
@@ -91,7 +91,7 @@ public final class Neighborhood implements Iterable<Point>
 	{
 		this.x = x;
 		this.y = y;
-		size = (int) goban.getBoardSize();
+		size = goban.getBoardSize();
 	}
 
 	@Override
